Adding error handling to git ops Updating Git form

+++ server/gitop.js
@@ -4,23 +4,49 @@ const UTF8 = 'utf-8';
 const DEFAULT_DIR = './';
 
 const getRemoteUrl = (rootDir = DEFAULT_DIR) => {
-  return execSync(`cd ${rootDir} && git config remote.origin.url`, { encoding: UTF8 });
+  try {
+    return execSync(`cd ${rootDir} && git config remote.origin.url`, { encoding: UTF8 });
+  } catch(e) {
+    return e.stderr;
+  }
 };
 
 const reset = (rootDir = DEFAULT_DIR) => {
-  return execSync(`cd ${rootDir} && git reset --hard`, { encoding: UTF8 });
+  try {
+    return execSync(`cd ${rootDir} && git reset --hard`, { encoding: UTF8 });
+  } catch(e) {
+    return e.stderr;
+  }
 };
 
 const deleteBranch = (rootDir = DEFAULT_DIR, branchname) => {
-  return execSync(`cd ${rootDir} && git branch -D ${branchname}`, { encoding: UTF8 });
+  try {
+    if(branchname) {
+      return execSync(`cd ${rootDir} && git branch -D ${branchname}`, { encoding: UTF8 });
+    }
+    return 'Missing branchname';
+  } catch(e) {
+    return e.stderr;
+  }
 };
 
 const selectBranch = (rootDir = DEFAULT_DIR, branchname) => {
-  return execSync(`cd ${rootDir} && git checkout ${branchname}`, { encoding: UTF8 });
+  try {
+    if(branchname) {
+      return execSync(`cd ${rootDir} && git checkout ${branchname}`, { encoding: UTF8 });
+    }
+    return 'Missing branchname';
+  } catch(e) {
+    return e.stderr;
+  }
 };
 
 const viewBranches = (rootDir = DEFAULT_DIR) => {
-  return execSync(`cd ${rootDir} && git branch`, { encoding: UTF8 });
+  try {
+    return execSync(`cd ${rootDir} && git branch`, { encoding: UTF8 });
+  } catch(e) {
+    return e.stderr;
+  }
 };
 
 const stash = (rootDir = DEFAULT_DIR, stashname) => {
@@ -31,14 +57,22 @@ const stash = (rootDir = DEFAULT_DIR, stashname) => {
 };
 
 const selectStash = (rootDir = DEFAULT_DIR, stashname) => {
-  if(stashname) {
-    return execSync(`cd ${rootDir} && git stash apply ${stashname}`, { encoding: UTF8 });
+  try {
+    if(stashname) {
+      return execSync(`cd ${rootDir} && git stash apply ${stashname}`, { encoding: UTF8 });
+    }
+    return execSync(`cd ${rootDir} && git stash apply`, { encoding: UTF8 });
+  } catch(e) {
+    return e.stderr;
   }
-  return execSync(`cd ${rootDir} && git stash apply`, { encoding: UTF8 });
 };
 
 const viewStash = (rootDir = DEFAULT_DIR) => {
-  return execSync(`cd ${rootDir} && git stash list`, { encoding: UTF8 });
+  try {
+    return execSync(`cd ${rootDir} && git stash list`, { encoding: UTF8 });
+  } catch(e) {
+    return e.stderr;
+  }
 };
 
 module.exports = {
